move getimagefromlocasotage func to utils
