+++ frontend/app/utils/localStorage/localStorage.ts
@@ -1,0 +1,13 @@
+export const getImageFromLocalStorage = (
+  key: string,
+  fallback: () => void
+): string => {
+  const image = localStorage.getItem(key)
+
+  if (!image) {
+    fallback()
+    return ''
+  }
+
+  return image
+}
